修复

+++ Interphone/app/src/main/java/com/interphone/connection/agreement/CmdPackage.java
@@ -49,7 +49,7 @@ public class CmdPackage {
   //数据错误
   public final static byte Cmd_type_error = 0x7f;
 
-  public static final byte CMD_TYPE_ACK = 0x7E;
+  public static final int CMD_TYPE_ACK = 0x7E;
 
   public static final int CMD_TYPE_ACK_CHANNEL_END = 0x7D;
